requirements.txt
# FAMIC Model Evaluation Requirements
# TODO: Update version numbers based on your specific requirements

# Core dependencies
torch>=1.9.0
numpy>=1.21.0

# Data handling
pandas>=1.3.0  # If using CSV datasets
scikit-learn>=0.24.0  # For evaluation metrics

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
PyYAML>=5.4.0

# Optional: For progress bars
tqdm>=4.62.0

# Optional: For Jupyter notebook support
# jupyter>=1.0.0
# ipykernel>=6.0.0

